Extend dev build to support yaml configuration
With this change one can run a dev build with yaml configuration.
Previously this was possible only with the `build-option` command
line flag. To use the yaml configuration you should add:
```
build:
  options:
    - dev
    # - debug #Just an example. This option is not supported yet.
```
For how to edit a language template to addopt the build-options
feature, please see `pytno3` template in the templates repository.

Signed-off-by: Ivana Yovcheva (VMware) <iyovcheva@vmware.com>